third_party/WebKit/LayoutTests/imported/wpt/html/dom/reflection-tabular-expected.txt

This is a testharness.js-based test.
PASS table.title: 77 tests
PASS table.lang: 77 tests
PASS table.dir: 151 tests
PASS table.className (<table class>): 77 tests
PASS table.hidden: 79 tests
PASS table.accessKey: 77 tests
PASS table.tabIndex: 10 tests
FAIL table.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS table.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL table.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS table.tabIndex: 36 tests
FAIL table.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS table.tabIndex: 2 tests
FAIL table.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS table.align: 77 tests
PASS table.border: 77 tests
PASS table.frame: 77 tests
PASS table.rules: 77 tests
PASS table.summary: 77 tests
PASS table.width: 77 tests
PASS table.bgColor: 77 tests
PASS table.cellPadding: 77 tests
PASS table.cellSpacing: 77 tests
PASS caption.title: 77 tests
PASS caption.lang: 77 tests
PASS caption.dir: 151 tests
PASS caption.className (<caption class>): 77 tests
PASS caption.hidden: 79 tests
PASS caption.accessKey: 77 tests
PASS caption.tabIndex: 10 tests
FAIL caption.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS caption.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL caption.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS caption.tabIndex: 36 tests
FAIL caption.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS caption.tabIndex: 2 tests
FAIL caption.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS caption.align: 77 tests
PASS colgroup.title: 77 tests
PASS colgroup.lang: 77 tests
PASS colgroup.dir: 151 tests
PASS colgroup.className (<colgroup class>): 77 tests
PASS colgroup.hidden: 79 tests
PASS colgroup.accessKey: 77 tests
PASS colgroup.tabIndex: 10 tests
FAIL colgroup.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS colgroup.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL colgroup.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS colgroup.tabIndex: 36 tests
FAIL colgroup.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS colgroup.tabIndex: 2 tests
FAIL colgroup.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS colgroup.span: 17 tests
FAIL colgroup.span: setAttribute() to 2147483648 followed by IDL get assert_equals: expected 1 but got 2147483648
PASS colgroup.span: setAttribute() to 4294967295 followed by getAttribute() 
FAIL colgroup.span: setAttribute() to 4294967295 followed by IDL get assert_equals: expected 1 but got 4294967295
PASS colgroup.span: 96 tests
FAIL colgroup.span: IDL set to 2147483648 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS colgroup.span: 2 tests
FAIL colgroup.span: IDL set to 4294967295 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS colgroup.span: IDL set to 4294967295 followed by IDL get 
PASS colgroup.align: 77 tests
PASS colgroup.ch (<colgroup char>): 77 tests
PASS colgroup.chOff (<colgroup charoff>): 77 tests
PASS colgroup.vAlign: 77 tests
PASS colgroup.width: 77 tests
PASS col.title: 77 tests
PASS col.lang: 77 tests
PASS col.dir: 151 tests
PASS col.className (<col class>): 77 tests
PASS col.hidden: 79 tests
PASS col.accessKey: 77 tests
PASS col.tabIndex: 10 tests
FAIL col.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS col.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL col.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS col.tabIndex: 36 tests
FAIL col.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS col.tabIndex: 2 tests
FAIL col.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS col.span: 17 tests
FAIL col.span: setAttribute() to 2147483648 followed by IDL get assert_equals: expected 1 but got 2147483648
PASS col.span: setAttribute() to 4294967295 followed by getAttribute() 
FAIL col.span: setAttribute() to 4294967295 followed by IDL get assert_equals: expected 1 but got 4294967295
PASS col.span: 96 tests
FAIL col.span: IDL set to 2147483648 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS col.span: 2 tests
FAIL col.span: IDL set to 4294967295 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS col.span: IDL set to 4294967295 followed by IDL get 
PASS col.align: 77 tests
PASS col.ch (<col char>): 77 tests
PASS col.chOff (<col charoff>): 77 tests
PASS col.vAlign: 77 tests
PASS col.width: 77 tests
PASS tbody.title: 77 tests
PASS tbody.lang: 77 tests
PASS tbody.dir: 151 tests
PASS tbody.className (<tbody class>): 77 tests
PASS tbody.hidden: 79 tests
PASS tbody.accessKey: 77 tests
PASS tbody.tabIndex: 10 tests
FAIL tbody.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS tbody.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL tbody.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS tbody.tabIndex: 36 tests
FAIL tbody.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS tbody.tabIndex: 2 tests
FAIL tbody.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS tbody.align: 77 tests
PASS tbody.ch (<tbody char>): 77 tests
PASS tbody.chOff (<tbody charoff>): 77 tests
PASS tbody.vAlign: 77 tests
PASS thead.title: 77 tests
PASS thead.lang: 77 tests
PASS thead.dir: 151 tests
PASS thead.className (<thead class>): 77 tests
PASS thead.hidden: 79 tests
PASS thead.accessKey: 77 tests
PASS thead.tabIndex: 10 tests
FAIL thead.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS thead.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL thead.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS thead.tabIndex: 36 tests
FAIL thead.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS thead.tabIndex: 2 tests
FAIL thead.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS thead.align: 77 tests
PASS thead.ch (<thead char>): 77 tests
PASS thead.chOff (<thead charoff>): 77 tests
PASS thead.vAlign: 77 tests
PASS tfoot.title: 77 tests
PASS tfoot.lang: 77 tests
PASS tfoot.dir: 151 tests
PASS tfoot.className (<tfoot class>): 77 tests
PASS tfoot.hidden: 79 tests
PASS tfoot.accessKey: 77 tests
PASS tfoot.tabIndex: 10 tests
FAIL tfoot.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS tfoot.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL tfoot.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS tfoot.tabIndex: 36 tests
FAIL tfoot.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS tfoot.tabIndex: 2 tests
FAIL tfoot.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS tfoot.align: 77 tests
PASS tfoot.ch (<tfoot char>): 77 tests
PASS tfoot.chOff (<tfoot charoff>): 77 tests
PASS tfoot.vAlign: 77 tests
PASS tr.title: 77 tests
PASS tr.lang: 77 tests
PASS tr.dir: 151 tests
PASS tr.className (<tr class>): 77 tests
PASS tr.hidden: 79 tests
PASS tr.accessKey: 77 tests
PASS tr.tabIndex: 10 tests
FAIL tr.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS tr.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL tr.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS tr.tabIndex: 36 tests
FAIL tr.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS tr.tabIndex: 2 tests
FAIL tr.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS tr.align: 77 tests
PASS tr.ch (<tr char>): 77 tests
PASS tr.chOff (<tr charoff>): 77 tests
PASS tr.vAlign: 77 tests
PASS tr.bgColor: 77 tests
PASS td.title: 77 tests
PASS td.lang: 77 tests
PASS td.dir: 151 tests
PASS td.className (<td class>): 77 tests
PASS td.hidden: 79 tests
PASS td.accessKey: 77 tests
PASS td.tabIndex: 10 tests
FAIL td.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS td.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL td.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS td.tabIndex: 36 tests
FAIL td.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS td.tabIndex: 2 tests
FAIL td.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS td.colSpan: 11 tests
FAIL td.colSpan: setAttribute() to 0 followed by IDL get assert_equals: expected 0 but got 1
PASS td.colSpan: 5 tests
FAIL td.colSpan: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 8190
PASS td.colSpan: setAttribute() to 2147483648 followed by getAttribute() 
FAIL td.colSpan: setAttribute() to 2147483648 followed by IDL get assert_equals: expected 1 but got 8190
PASS td.colSpan: setAttribute() to 4294967295 followed by getAttribute() 
FAIL td.colSpan: setAttribute() to 4294967295 followed by IDL get assert_equals: expected 1 but got 8190
PASS td.colSpan: 7 tests
FAIL td.colSpan: setAttribute() to "-0" followed by IDL get assert_equals: expected 0 but got 1
PASS td.colSpan: setAttribute() to "0" followed by getAttribute() 
FAIL td.colSpan: setAttribute() to "0" followed by IDL get assert_equals: expected 0 but got 1
PASS td.colSpan: 78 tests
FAIL td.colSpan: IDL set to 0 followed by IDL get assert_equals: expected 0 but got 1
PASS td.colSpan: 8 tests
FAIL td.colSpan: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 8190
PASS td.colSpan: 2 tests
FAIL td.colSpan: IDL set to "-0" followed by IDL get assert_equals: expected 0 but got 1
PASS td.colSpan: IDL set to 2147483648 should not throw 
FAIL td.colSpan: IDL set to 2147483648 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS td.colSpan: 2 tests
FAIL td.colSpan: IDL set to 4294967295 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS td.colSpan: IDL set to 4294967295 followed by IDL get 
PASS td.rowSpan: 11 tests
FAIL td.rowSpan: setAttribute() to 0 followed by IDL get assert_equals: expected 0 but got 1
PASS td.rowSpan: 5 tests
FAIL td.rowSpan: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 65534
PASS td.rowSpan: setAttribute() to 2147483648 followed by getAttribute() 
FAIL td.rowSpan: setAttribute() to 2147483648 followed by IDL get assert_equals: expected 1 but got 65534
PASS td.rowSpan: setAttribute() to 4294967295 followed by getAttribute() 
FAIL td.rowSpan: setAttribute() to 4294967295 followed by IDL get assert_equals: expected 1 but got 65534
PASS td.rowSpan: 7 tests
FAIL td.rowSpan: setAttribute() to "-0" followed by IDL get assert_equals: expected 0 but got 1
PASS td.rowSpan: setAttribute() to "0" followed by getAttribute() 
FAIL td.rowSpan: setAttribute() to "0" followed by IDL get assert_equals: expected 0 but got 1
PASS td.rowSpan: 78 tests
FAIL td.rowSpan: IDL set to 0 followed by IDL get assert_equals: expected 0 but got 1
PASS td.rowSpan: 8 tests
FAIL td.rowSpan: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 65534
PASS td.rowSpan: 2 tests
FAIL td.rowSpan: IDL set to "-0" followed by IDL get assert_equals: expected 0 but got 1
PASS td.rowSpan: IDL set to 2147483648 should not throw 
FAIL td.rowSpan: IDL set to 2147483648 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS td.rowSpan: 2 tests
FAIL td.rowSpan: IDL set to 4294967295 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS td.rowSpan: IDL set to 4294967295 followed by IDL get 
PASS td.scope: 176 tests
PASS td.abbr: 77 tests
PASS td.align: 77 tests
PASS td.axis: 77 tests
PASS td.height: 77 tests
PASS td.width: 77 tests
PASS td.ch (<td char>): 77 tests
PASS td.chOff (<td charoff>): 77 tests
PASS td.noWrap: 79 tests
PASS td.vAlign: 77 tests
PASS td.bgColor: 77 tests
PASS th.title: 77 tests
PASS th.lang: 77 tests
PASS th.dir: 151 tests
PASS th.className (<th class>): 77 tests
PASS th.hidden: 79 tests
PASS th.accessKey: 77 tests
PASS th.tabIndex: 10 tests
FAIL th.tabIndex: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS th.tabIndex: setAttribute() to -2147483648 followed by getAttribute() 
FAIL th.tabIndex: setAttribute() to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS th.tabIndex: 36 tests
FAIL th.tabIndex: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 32767
PASS th.tabIndex: 2 tests
FAIL th.tabIndex: IDL set to -2147483648 followed by IDL get assert_equals: expected -2147483648 but got -32768
PASS th.colSpan: 11 tests
FAIL th.colSpan: setAttribute() to 0 followed by IDL get assert_equals: expected 0 but got 1
PASS th.colSpan: 5 tests
FAIL th.colSpan: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 8190
PASS th.colSpan: setAttribute() to 2147483648 followed by getAttribute() 
FAIL th.colSpan: setAttribute() to 2147483648 followed by IDL get assert_equals: expected 1 but got 8190
PASS th.colSpan: setAttribute() to 4294967295 followed by getAttribute() 
FAIL th.colSpan: setAttribute() to 4294967295 followed by IDL get assert_equals: expected 1 but got 8190
PASS th.colSpan: 7 tests
FAIL th.colSpan: setAttribute() to "-0" followed by IDL get assert_equals: expected 0 but got 1
PASS th.colSpan: setAttribute() to "0" followed by getAttribute() 
FAIL th.colSpan: setAttribute() to "0" followed by IDL get assert_equals: expected 0 but got 1
PASS th.colSpan: 78 tests
FAIL th.colSpan: IDL set to 0 followed by IDL get assert_equals: expected 0 but got 1
PASS th.colSpan: 8 tests
FAIL th.colSpan: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 8190
PASS th.colSpan: 2 tests
FAIL th.colSpan: IDL set to "-0" followed by IDL get assert_equals: expected 0 but got 1
PASS th.colSpan: IDL set to 2147483648 should not throw 
FAIL th.colSpan: IDL set to 2147483648 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS th.colSpan: 2 tests
FAIL th.colSpan: IDL set to 4294967295 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS th.colSpan: IDL set to 4294967295 followed by IDL get 
PASS th.rowSpan: 11 tests
FAIL th.rowSpan: setAttribute() to 0 followed by IDL get assert_equals: expected 0 but got 1
PASS th.rowSpan: 5 tests
FAIL th.rowSpan: setAttribute() to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 65534
PASS th.rowSpan: setAttribute() to 2147483648 followed by getAttribute() 
FAIL th.rowSpan: setAttribute() to 2147483648 followed by IDL get assert_equals: expected 1 but got 65534
PASS th.rowSpan: setAttribute() to 4294967295 followed by getAttribute() 
FAIL th.rowSpan: setAttribute() to 4294967295 followed by IDL get assert_equals: expected 1 but got 65534
PASS th.rowSpan: 7 tests
FAIL th.rowSpan: setAttribute() to "-0" followed by IDL get assert_equals: expected 0 but got 1
PASS th.rowSpan: setAttribute() to "0" followed by getAttribute() 
FAIL th.rowSpan: setAttribute() to "0" followed by IDL get assert_equals: expected 0 but got 1
PASS th.rowSpan: 78 tests
FAIL th.rowSpan: IDL set to 0 followed by IDL get assert_equals: expected 0 but got 1
PASS th.rowSpan: 8 tests
FAIL th.rowSpan: IDL set to 2147483647 followed by IDL get assert_equals: expected 2147483647 but got 65534
PASS th.rowSpan: 2 tests
FAIL th.rowSpan: IDL set to "-0" followed by IDL get assert_equals: expected 0 but got 1
PASS th.rowSpan: IDL set to 2147483648 should not throw 
FAIL th.rowSpan: IDL set to 2147483648 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS th.rowSpan: 2 tests
FAIL th.rowSpan: IDL set to 4294967295 followed by getAttribute() assert_equals: expected "1" but got "0"
PASS th.rowSpan: IDL set to 4294967295 followed by IDL get 
PASS th.scope: 176 tests
PASS th.abbr: 77 tests
PASS th.align: 77 tests
PASS th.axis: 77 tests
PASS th.height: 77 tests
PASS th.width: 77 tests
PASS th.ch (<th char>): 77 tests
PASS th.chOff (<th charoff>): 77 tests
PASS th.noWrap: 79 tests
PASS th.vAlign: 77 tests
PASS th.bgColor: 77 tests
Harness: the test ran to completion.
